improvements

+++ client/src/components/watchlist-card/watchlist-card-item/WatchlistDeleteButton.tsx
@@ -11,15 +11,28 @@ import { useToast } from '@/hooks/use-toast';
 import { useRouter } from 'next/navigation';
 import { useApi } from '@/hooks/use-api';
 import { useWatchlist } from '@/providers/WatchlistProvider';
+import { useConfirmDialog } from '@/components/confirm-dialog/ConfirmDialogProvider';
 
 const WatchlistDeleteButton = ({}) => {
   const { toast } = useToast();
   const router = useRouter();
   const api = useApi();
+  const showConfirmDialog = useConfirmDialog();
+
   const { title, id: watchlistId } = useWatchlist();
 
   const handleDelete = async () => {
     try {
+      const confirmed = await showConfirmDialog({
+        title: `Are you sure you want to delete watchlist ${title}?`,
+        description:
+          'Deleting watchlist will delete watchlist for you and all watchlist users. Deleting cannot be undone.'
+      });
+
+      if (!confirmed) {
+        return;
+      }
+
       await api.delete(`/watchlists/${watchlistId}`);
 
       toast({
